Refactor component + remove Formik dep
